[REF] deltatech_uom_unece: merge the EDI bridge into the single module
The bridge was split off so the main module would stay installable where no
UBL or CII invoice is ever sent. That case barely exists: `account_edi_ubl_cii`
is itself `auto_install`, so it is present on any database with accounting, and
the bridge came along with it. Verified by installing the main module alone on a
clean database — `account_edi_ubl_cii` and the bridge both ended up installed
without being asked for.

So the split bought structural tidiness and cost a second module, a second
manifest and a second banner. Fold the `account.edi.common` override back in and
depend on `account_edi_ubl_cii` directly.

Note that the override is not optional: `account_edi_ubl_cii` does not call
`uom.uom._get_unece_code()`, it keeps a second copy of the logic of its own.
Overriding only the method on `uom.uom` fixes eTransport and silently leaves
invoices unchanged.
