+++ deltatech_uom_unece/models/__init__.py
@@ -1,2 +1,3 @@
 from . import uom_unece_code
 from . import uom_uom
+from . import account_edi_common
