+++ deltatech_uom_unece/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_uom_unece
+from . import test_uom_unece_edi
