--- deltatech_uom_unece_edi/__init__.py
@@ -1,1 +1,0 @@
-from . import models
