--- deltatech_uom_unece_edi/__manifest__.py
@@ -1,19 +1,0 @@
-# ©  2026 Deltatech
-#              Dorin Hongu <dhongu(@)gmail(.)com
-# See README.rst file on addons root folder for license details
-
-{
-    "images": ["static/description/main_screenshot.png"],
-    "name": "Deltatech UoM UNECE Codes - EDI Invoices",
-    "summary": "Use the UNECE code set on the unit in UBL/CII electronic invoices",
-    "version": "19.0.1.0.0",
-    "author": "Terrabit,Dorin Hongu",
-    "website": "https://www.terrabit.ro",
-    "category": "Administration",
-    "depends": ["deltatech_uom_unece", "account_edi_ubl_cii"],
-    "license": "OPL-1",
-    "installable": True,
-    "auto_install": True,
-    "development_status": "Beta",
-    "maintainers": ["dhongu"],
-}
