--- deltatech_uom_unece_edi/models/__init__.py
@@ -1,1 +1,0 @@
-from . import account_edi_common
