--- deltatech_uom_unece_edi/pyproject.toml
@@ -1,8 +1,0 @@
-[build-system]
-requires = [
-    "whool",
-]
-build-backend = "whool.buildapi"
-
-[project]
-name = "odoo-addon-deltatech-uom-unece-edi"
