--- deltatech_uom_unece_edi/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_uom_unece_edi
